requirements.txt
# AI Agent Dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.28.0
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
typing-extensions>=4.0.0

# Weather and Email Dependencies
# openweathermap-api>=0.1.0  # Removed - using direct API calls
# smtplib and email are built-in Python libraries

# GCP Dependencies
google-cloud-scheduler>=2.0.0
google-cloud-functions>=1.0.0
google-cloud-logging>=3.0.0

# Optional: For more advanced features
# chromadb>=0.4.0  # For vector storage
# tiktoken>=0.5.0   # For token counting
# numpy>=1.24.0     # For numerical operations
# pandas>=2.0.0     # For data manipulation 